starship-ai/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
argcomplete==3.6.0
asgi-correlation-id==4.3.4
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.2.0
boto3==1.40.36
botocore==1.40.36
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
decorator==5.2.1
dependency-groups==1.3.0
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
jmespath==1.0.1
jsonpatch==1.33
jsonpath-rw==1.4.0
jsonpointer==3.0.0
mypy-extensions==1.0.0
nox==2025.2.9
packaging==24.2
parse==1.20.2
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
pydantic==2.10.6
pydantic_core==2.27.2
pyproject-api==1.9.0
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-ilorest-library==6.0.0.0
python-mock==0.0.3
PyYAML==6.0.2
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.44.0
tomli==2.2.1
tomlkit==0.13.2
tox==4.24.2
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.33.0
virtualenv==20.29.3
vox==0.0.1
xmltodict==0.14.2
zipp==3.20.2
streamlit>=1.28.0
python-dotenv>=1.0.0
